change array type form int to float

+++ Tools/PIDML/qaPidML.cxx
@@ -321,7 +321,7 @@ struct pidml {
   template <std::size_t i, typename T>
   void pidML(const T& track, const int pdgCodeMC)
   {
-    int pidLogits[3] = {model211.applyModel(track), model2212.applyModel(track), model321.applyModel(track)};
+    float pidLogits[3] = {int(model211.applyModel(track)), model2212.applyModel(track), model321.applyModel(track)};
     int pid = pidParticle(pidLogits);
     if (pid == particlesPdgCode[i]) {
       if (pdgCodeMC == particlesPdgCode[i]) {
